Move Secret and VolumeWithMount.
