+++ scheduler/src/main/scala/mesosphere/marathon/state/Secret.scala
@@ -1,0 +1,6 @@
+package mesosphere.marathon
+package state
+
+import mesosphere.marathon.plugin
+
+case class Secret(source: String) extends plugin.Secret
